kube-aws: reduce complexity of cloudformation template
Figure out which ami to use in kube-aws application logic, and template into cloudformation template on `kube-aws up` templating pass. Now consistent with how everything else is done.

Cloudformation no longer is relied on to infer availability zone if it is not supplied (availability zone must now be supplied by the user). For EBS volume attachmenents (a future PR), it is critical that the availability zone be supplied explicitly by the user.